requirements.txt
flask
bs4
requests
google-cloud-storage==2.7.0
